% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfMatrixBuild.R
\name{Build_post_summary}
\alias{Build_post_summary}
\title{Build count matrix for input samples}
\usage{
Build_post_summary(
  input,
  output = NA,
  typefile = NA,
  minsample = 30,
  multicore = FALSE,
  TCGA = F,
  ICGC = F
)
}
\arguments{
\item{input}{ccf file foder path}

\item{output}{output path}

\item{typefile}{type with sample-type mapping}
}
\value{
a data frame containing the summary for all samples in the ccf files
}
\description{
Build count matrix for input samples
}
