% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfMatrixBuild.R
\name{ccfMatBuild}
\alias{ccfMatBuild}
\title{Build CCF Matrix for all samples within region 0-upper}
\usage{
ccfMatBuild(
  samplelist,
  upper,
  input_folder,
  binwidth = 0.01,
  add_samplename = FALSE,
  plot = FALSE,
  random = FALSE
)
}
\arguments{
\item{samplelist}{samples' nam+e}

\item{upper}{set CCF upper bound}

\item{input_folder}{folder path stores ccf files}

\item{binwidth}{ccf bin width}

\item{add_samplename}{specify whether to add samplename variables}

\item{plot}{whether to plot mutation distribution for samples}

\item{random}{whether to output randomized ccf count matrix}

\item{genelist}{specify whether to only count for certain genes}
}
\value{
A list containing ccfBandCountsMat,ccfBandCountsMat.random,ccfBandFractionMat and ccfBandFractionMat.random
}
\description{
Build CCF Matrix for all samples within region 0-upper
}
