% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sig_nmf.R
\name{combine_sig_nmf}
\alias{combine_sig_nmf}
\title{Combine signatures from different cancer type together}
\usage{
combine_sig_nmf(input_folder, cancertype)
}
\arguments{
\item{input_folder}{folder stores signatures for each type}

\item{cancertype}{cancer type list}
}
\value{
matrix combining all signature matrix
}
\description{
Combine signatures from different cancer type together
}
