% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Association_heatmap.r
\name{cor_evoSig}
\alias{cor_evoSig}
\title{compute correlation for a single measure with evolutionary dynamics signatures}
\usage{
cor_evoSig(file, varcol, n_sig, output = "r")
}
\arguments{
\item{file}{exposure file merged with measures of interest}

\item{varcol}{number of evolutionary dynamics signatures}

\item{n_sig}{input dataframe containing all samples with corresponding types}

\item{output}{specify ouput correlation value, p value or sample size}
}
\value{
depends on output choice
}
\description{
compute correlation for a single measure with evolutionary dynamics signatures
}
