% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Association_heatmap.r
\name{Cor_heatmap}
\alias{Cor_heatmap}
\title{Correlation heatmap plotting}
\usage{
Cor_heatmap(
  file,
  var_index,
  output = "NA",
  n_sig,
  minsample,
  width = 12,
  height = 30,
  facet
)
}
\arguments{
\item{file}{input exposure file with measures}

\item{output}{output folder path}

\item{n_sig}{number of signatures}

\item{minsample}{minimun samples required for computing correlation}

\item{width}{width of output plot}

\item{height}{height of output plot}

\item{facet}{specify facet}
}
\value{
return heatmap for correlation between measures and evolutionary signature
}
\description{
Correlation heatmap plotting
}
