% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Association_heatmap.r
\name{Cor_heatmap_colon}
\alias{Cor_heatmap_colon}
\title{Correlation heatmap plotting}
\usage{
Cor_heatmap_colon(
  file,
  var_index,
  immune_measures_mapping = immune_measures_mapping,
  output = "NA",
  n_sig,
  immune = TRUE,
  var = TRUE,
  group = "cancertype"
)
}
\arguments{
\item{file}{input exposure file with measures}

\item{output}{output folder path}

\item{n_sig}{number of signatures}

\item{width}{width of output plot}

\item{height}{height of output plot}

\item{facet}{specify facet}

\item{minsample}{minimun samples required for computing correlation}
}
\value{
return heatmap for correlation between measures and evolutionary signature
}
\description{
Correlation heatmap plotting
}
