% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Association_heatmap.r
\name{corr_spearman_type}
\alias{corr_spearman_type}
\title{Basic correlation function for each cell}
\usage{
corr_spearman_type(xx, n_sig, varcol, output)
}
\arguments{
\item{xx}{exposure file merged with measures of interest}

\item{n_sig}{number of evolutionary dynamics signatures}

\item{varcol}{input dataframe containing all samples with corresponding types}

\item{output}{specify ouput correlation value, p value or sample size}
}
\value{
depends on output choice
}
\description{
Basic correlation function for each cell
}
