% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TypeMatrixBuild.R
\name{CountMatBuild}
\alias{CountMatBuild}
\title{Build CCF Matrix for all samples within region 0-upper}
\usage{
CountMatBuild(samplelist, upper, input_folder, genelist = NA)
}
\arguments{
\item{samplelist}{samples' name}

\item{upper}{set CCF upper bound}

\item{input_folder}{folder path stores ccf files}

\item{genelist}{specify whether to only count for certain genes}
}
\value{
a data frame containing the summary for all samples in the ccf files
}
\description{
Build CCF Matrix for all samples within region 0-upper
}
