% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrap_up.R
\name{EvoSig_input}
\alias{EvoSig_input}
\title{Produce input ccf count matrix for rank estimate}
\usage{
EvoSig_input(
  Ccube_folder,
  output,
  ccfMatBuild = F,
  post_summary = NA,
  postSummaryBuild = F,
  TCGA = F,
  ICGC = F,
  ccfupper = 1,
  minsample = 30
)
}
\arguments{
\item{Ccube_folder}{Ccube folder}

\item{output}{output folder path}

\item{ccfMatBuild}{Specify whether construct ccf count matrix or not}

\item{post_summary}{specify post summary file path if choose to load from local file}

\item{postSummaryBuild}{Specify whether construct post summary from scratch or not}

\item{ccfupper}{CCF upper bound}

\item{minsample}{set minimum samples needed for a single cancer type}
}
\value{
CCF matrix for each cancer type
}
\description{
Produce input ccf count matrix for rank estimate
}
