% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_sig.R
\name{Extract_sig}
\alias{Extract_sig}
\title{Perform NMF with consensus signature}
\usage{
Extract_sig(ccfMat, consensus_sig, samplelist, output = NA)
}
\arguments{
\item{ccfMat}{ccf matrix for all samples}

\item{consensus_sig}{consensus signature matrix}

\item{samplelist}{input sample list}

\item{output}{output folder path}
}
\value{
exposure
}
\description{
Perform NMF with consensus signature
}
