% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_cluster_test.R
\name{hc_cluster_test ccfMat ccf matrix for all samples}
\alias{hc_cluster_test ccfMat ccf matrix for all samples}
\alias{hc_cluster_test}
\title{Perform Hierarchical clustering number estimate}
\usage{
hc_cluster_test(data, methods, distance, min = 2, max = 10)
}
\arguments{
\item{data}{combined signatures for all cancer types}

\item{methods}{clustering method}

\item{distance}{distance funciton}

\item{min}{minimum clustering number}

\item{max}{maximum clustering number}
}
\value{
exposure
}
\description{
Perform Hierarchical clustering number estimate
}
