% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_consensus.R
\name{hc_combine_sig Combine signatures and estimate number of clusters}
\alias{hc_combine_sig Combine signatures and estimate number of clusters}
\alias{hc_combine_sig}
\title{Combine signatures and estimate number of clusters}
\usage{
hc_combine_sig(
  input_folder,
  output_folder,
  cancertype,
  min.nc = 3,
  max.nc = 10,
  method = "ward.D2",
  index = "hubert",
  distance = "euclidean"
)
}
\arguments{
\item{input_folder}{input folder}

\item{output_folder}{output folder path}

\item{cancertype}{cancer types}

\item{min.nc}{minimum cluster}

\item{max.nc}{maximum cluster}

\item{distance}{distance function}

\item{ccfMatrix}{ccf matrix for all samples}
}
\value{
list(combine_sig,cluster)
}
\description{
Combine signatures and estimate number of clusters
}
