% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_consensus.R
\name{hc_consensus ccfMat ccf matrix for all samples}
\alias{hc_consensus ccfMat ccf matrix for all samples}
\alias{hc_consensus_sig}
\alias{sig_assignment}
\title{Plotting for HC results}
\usage{
hc_consensus_sig(
  combine_sig,
  cluster,
  output_folder,
  method = "ward.D2",
  index = "hubert",
  distance = "euclidean"
)

sig_assignment(signature, ccfMatrix, output = NA)
}
\arguments{
\item{combine_sig}{combined signatures for all cancer types}

\item{cluster}{clustering number}

\item{output_folder}{output folder path}

\item{distance}{distance function}

\item{signature}{signature}

\item{ccfMatrix}{ccfmatrix}

\item{output}{output folder path}
}
\value{
consensus_sig

exposure
}
\description{
Plotting for HC results

sig_assignment
}
