% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Association_heatmap.r
\name{heatmap_plot_for_single_measure_across_types}
\alias{heatmap_plot_for_single_measure_across_types}
\title{Correlation of a single measure with evolutionary signatures across cancer types}
\usage{
heatmap_plot_for_single_measure_across_types(
  file,
  varcol,
  n_sig,
  keep_pl0.05 = FALSE,
  minsample = minsample,
  Type
)
}
\arguments{
\item{file}{input file}

\item{varcol}{columns index of variable of interest}

\item{n_sig}{number of signatures}

\item{keep_pl0.05}{whether to keep correaltion with p>0.05}

\item{minsample}{the minimun samples}

\item{Type}{Specify whether to compute correlation for each cancer type}
}
\value{
a data frame containing the summary for all samples in the ccf files
}
\description{
Correlation of a single measure with evolutionary signatures across cancer types
}
