% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Association_heatmap.r
\name{heatmap_plot_multi_groups}
\alias{heatmap_plot_multi_groups}
\alias{cor_table_multi_variable_groups}
\title{Obtain correlation table with r,p,n between signature exposures and measures of interest (across group)}
\usage{
cor_table_multi_variable_groups(file, var_idx, n_sig, group = NA)
}
\arguments{
\item{file}{exposure file with measures of interest}

\item{n_sig}{number of signatures}

\item{group}{specify group variables}

\item{varcol}{index of columns of measures}
}
\value{
return correlation table with r,p,n between signature exposures and measures of interest (across group)
}
\description{
Obtain correlation table with r,p,n between signature exposures and measures of interest (across group)
}
