% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Association_heatmap.r
\name{heatmap_plot_multi_variable}
\alias{heatmap_plot_multi_variable}
\title{Plot heatmap for correlation between measures and evolutionary signature}
\usage{
heatmap_plot_multi_variable(file, varcol, n_sig, Type = TRUE, minsample = 10)
}
\arguments{
\item{file}{exposure file with measures of interest}

\item{varcol}{index of columns of measures}

\item{n_sig}{number of signatures}

\item{Type}{Specify compute correlation for each cancer type or for all}

\item{minsample}{minimun samples required for computing correlation}
}
\value{
return heatmap for correlation between measures and evolutionary signature
}
\description{
Plot heatmap for correlation between measures and evolutionary signature
}
