% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_landscape_measure.R
\docType{data}
\name{immune_landscape_measure}
\alias{immune_landscape_measure}
\title{Immune landscape measures dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11080 rows and 64 columns.
}
\usage{
immune_landscape_measure
}
\description{
Immune landscape measures dataset
}
\keyword{datasets}
