% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{load_ccf}
\alias{load_ccf}
\title{Load multiple format ccf file}
\usage{
load_ccf(samplename, input)
}
\arguments{
\item{samplename}{cancer type}

\item{input}{}
}
\value{
ssm
}
\description{
Load multiple format ccf file
}
