% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{multi.dir.create}
\alias{multi.dir.create}
\alias{multi_dir_create}
\title{create multiple dir}
\usage{
multi_dir_create(dirlist)
}
\arguments{
\item{list}{list of directory}
}
\value{
create multiple directory
}
\description{
create multiple dir
}
