% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_sig_all_plot.R
\name{nmf_sig_all_plot}
\alias{nmf_sig_all_plot}
\title{Perform nmf for multiple cancer types}
\usage{
nmf_sig_all_plot(
  input_folder,
  output,
  rank_summary,
  MatType = "fraction",
  nrun = 100
)
}
\arguments{
\item{input_folder}{ccf file path}

\item{output}{output file path}

\item{rank_summary}{rank summary file}
}
\value{
save nmf results in output folder and plot signature for all cancer types
}
\description{
Perform nmf for multiple cancer types
}
