% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_sig_all_plot.R
\name{nmf_sig_plot_type}
\alias{nmf_sig_plot_type}
\title{Perform nmf for a cancer type}
\usage{
nmf_sig_plot_type(type, MatType = "fraction", input_folder, output, rank)
}
\arguments{
\item{type}{cancer type}

\item{MatType}{matrix type}

\item{input_folder}{ccf file path}

\item{output}{output file path}

\item{rank}{rank}
}
\value{
save nmf results in output folder and plot signature for this type
}
\description{
Perform nmf for a cancer type
}
