% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{p_scatter}
\alias{p_scatter}
\title{Plot scatterplot for specific measure with evo sig and subtype - 20200915}
\usage{
p_scatter(df, vb, var, sig, facet)
}
\arguments{
\item{df}{exposure file merged with measures of interest}

\item{vb}{Column indexs of all measures in association heatmap in df}

\item{sig}{Column indexs of evo signature}

\item{facet}{name of facet variable consistent with association heatmap}
}
\value{
Scatterplots
}
\description{
Plot scatterplot for specific measure with evo sig and subtype - 20200915
}
