% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{ParseSnvCnaPcawgFormat}
\alias{ParseSnvCnaPcawgFormat}
\title{Merge ssm and cna files}
\usage{
ParseSnvCnaPcawgFormat(ssm, cna)
}
\arguments{
\item{ssm}{ssm}

\item{cna}{cna}
}
\description{
Merge ssm and cna files
}
