% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_estimate_plot.R
\name{rank_estimate_plot}
\alias{rank_estimate_plot}
\title{Plotting rank estimate}
\usage{
rank_estimate_plot(outputFolder, rankfilepath)
}
\arguments{
\item{outputFolder}{folder stores rank estimate files}

\item{rankfilepath}{rank file path to output}
}
\value{
save nmf results in output folder and plot signature for all cancer types
}
\description{
Plotting rank estimate
}
