% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_sig_all_plot.R
\name{sig_plot}
\alias{sig_plot}
\title{Plot signature matrix}
\usage{
sig_plot(sig, tag = NULL, col = NA, theme = "grey")
}
\arguments{
\item{sig}{signature matrix}
}
\value{
mydata mutation data frame
}
\description{
Plot signature matrix
}
