% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNV_exposure.R
\docType{data}
\name{SNV_exposure}
\alias{SNV_exposure}
\title{SNV_exposure dataset}
\format{An object of class \code{data.frame} with 10179 rows and 31 columns.}
\usage{
SNV_exposure
}
\description{
SNV_exposure dataset
}
\keyword{datasets}
