% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAtypefile.R
\docType{data}
\name{TCGAtypefile}
\alias{TCGAtypefile}
\title{TCGAtypefile dataset}
\format{
An object of class \code{data.frame} with 10186 rows and 3 columns.
}
\usage{
TCGAtypefile
}
\description{
TCGAtypefile dataset
}
\keyword{datasets}
