% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TypeMatrixBuild.R
\name{TypeMatrixBuild}
\alias{TypeMatrixBuild}
\title{Build count matrix for input samples per cancer type}
\usage{
TypeMatrixBuild(
  post_summary,
  input_folder,
  output,
  ccfupper = 1,
  RankEstimateType = "fraction"
)
}
\arguments{
\item{post_summary}{samples' name}

\item{input_folder}{folder path stores ccf files}

\item{output}{output folder path}

\item{ccfupper}{set CCF upper bound}

\item{RankEstimateType}{output rank estimate matirx format}

\item{genelist}{specify whether to only count for certain genes}
}
\value{
CCF matrix for each cancer type
}
\description{
Build count matrix for input samples per cancer type
}
