% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfMatrixBuild.R
\name{ccfMatBuild}
\alias{ccfMatBuild}
\title{Build CCF Matrix for all samples within region 0-upper}
\usage{
ccfMatBuild(
  samplelist,
  upper,
  input_folder,
  genelist = NA,
  add_samplename = FALSE
)
}
\arguments{
\item{samplelist}{samples' nam+e}

\item{upper}{set CCF upper bound}

\item{input_folder}{folder path stores ccf files}

\item{genelist}{specify whether to only count for certain genes}

\item{add_samplename}{specify whether to add samplename variables}
}
\value{
A list containing ccfBandCountsMat,ccfBandCountsMat.random,ccfBandFractionMat and ccfBandFractionMat.random
}
\description{
Build CCF Matrix for all samples within region 0-upper
}
