% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{cor_facet}
\alias{cor_facet}
\title{Correlation calculation function by group and plot association heatmap - 20200915}
\usage{
cor_facet(
  df,
  va,
  vb,
  facet,
  heatmap = FALSE,
  title = "",
  empty_row_delete = FALSE,
  flip = FALSE,
  keep_all = TRUE
)
}
\arguments{
\item{df}{exposure file merged with measures of interest}

\item{va}{Column indexs of variables A in df}

\item{vb}{Column indexs of variables B in df}

\item{facet}{name of group variable}

\item{heatmap}{specify whether plot association heatmap}

\item{title}{Title for association heatmap}
}
\value{
A table including correlation r,p,n between variables A and variables B during group variable and association heatmap (optional)
}
\description{
Correlation calculation function by group and plot association heatmap - 20200915
}
