% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_consensus.R
\name{hc_consensus ccfMat ccf matrix for all samples}
\alias{hc_consensus ccfMat ccf matrix for all samples}
\alias{hc_consensus}
\title{Plotting for HC results}
\usage{
hc_consensus(combine_sig, cluster, output, ccfMatrix, distance = "euclidean")
}
\arguments{
\item{combine_sig}{combined signatures for all cancer types}

\item{cluster}{clustering number}

\item{output}{output folder path}

\item{ccfMatrix}{ccf matrix for all samples}

\item{distance}{distance function}
}
\value{
exposure
}
\description{
Plotting for HC results
}
